Quadruple video! Capture vids from each of the four sides.

+++ utils/video_export.py
@@ -5,6 +5,7 @@ All export functions are no-ops unless export has been enabled in config.py
 StackOverflow: https://stackoverflow.com/a/62434934
 Documentation of MoviePy: https://zulko.github.io/moviepy/index.html
 """
+from dataclasses import dataclass
 import math
 import os
 
@@ -17,6 +18,12 @@ import epiboly_globals as g
 import utils.epiboly_utils as epu
 import utils.tf_utils as tfu
 
+@dataclass
+class CameraData:
+    center: tf.fVector3
+    rotation: tf.fQuaternion
+    zoom: float
+
 def final_result_screenshots() -> None:
     """If enabled, capture still images from multiple angles.
 
@@ -26,27 +33,27 @@ def final_result_screenshots() -> None:
     if not cfg.windowed_mode and screenshot_export_enabled():
         tf.system.camera_view_front()
         tf.system.camera_zoom_to(-12)
-        save_screenshot("Front", show_timestep=False)
+        _save_screenshot("Front", show_timestep=False)
         
         tf.system.camera_view_left()
         tf.system.camera_zoom_to(-12)
-        save_screenshot("Left", show_timestep=False)
+        _save_screenshot("Left", show_timestep=False)
         
         tf.system.camera_view_back()
         tf.system.camera_zoom_to(-12)
-        save_screenshot("Back", show_timestep=False)
+        _save_screenshot("Back", show_timestep=False)
         
         tf.system.camera_view_right()
         tf.system.camera_zoom_to(-12)
-        save_screenshot("Right", show_timestep=False)
+        _save_screenshot("Right", show_timestep=False)
 
         tf.system.camera_view_top()
         tf.system.camera_zoom_to(-12)
-        save_screenshot("Top", show_timestep=False)
+        _save_screenshot("Top", show_timestep=False)
 
         tf.system.camera_view_bottom()
         tf.system.camera_zoom_to(-12)
-        save_screenshot("Bottom", show_timestep=False)
+        _save_screenshot("Bottom", show_timestep=False)
 
 _rotation_started: bool = False
 _rotation_finished: bool = False
@@ -69,6 +76,9 @@ def _test_when_to_finish_rotation() -> None:
     is a bit weird; I selected angle() as the relevant property, and dropping below pi as the relevant behavior,
     based on empirical testing, but it may be very specific to this particular rotation. Seems like this may be brittle.
     (In fact, a small refactor – no algorithmic change – necessitated an adjustment in the tolerance, inexplicably.)
+    (And, now that I've added the quadruple video, with 4 screenshots per export, one from each side, "Front" is quite
+    intentionally done last, otherwise this wouldn't work. This is called after all four screenshots are done,
+    and the criterion is specific to the rotation on that side.)
     """
     global _rotation_finished
     
@@ -76,10 +86,17 @@ def _test_when_to_finish_rotation() -> None:
     # pole), and then continues to drop lower if you rotate past Bottom and up the other side. So we want to stop
     # when angle < π. However, it does not hit exactly. At very bottom, value is slightly greater than π;
     # stopping at π thus overshoots by one rotation increment. So add a small tolerance to stop in the right place.
-    target_camera_angle: float = math.pi + 0.04
+    # Furthermore, the timing of when modified camera data becomes readable is different for windowed mode
+    # (maybe because I'm not moving the camera around as much, or maybe due to internal TF issues); either way,
+    # in windowed mode, the angle we read is for the previous position so we need to stop rotating one
+    # timestep earlier, so just make the tolerance bigger.
+    tolerance: float = 0.04 if cfg.windowed_mode else 0.01
+    target_camera_angle: float = math.pi + tolerance
     
     quat: tf.fQuaternion = tf.system.camera_rotation()
+    # print(f"Target  = {target_camera_angle}\ncurrent = {quat.angle()}")
     if quat.angle() < target_camera_angle:
+        # print("Setting rotation finished!")
         _rotation_finished = True
 
 def _automated_camera_rotate() -> None:
@@ -89,16 +106,12 @@ def _automated_camera_rotate() -> None:
     # if cfg.windowed_mode:
     #     return
     
-    if not _rotation_started:
-        _test_when_to_start_rotation()
-    
     if _rotation_started and not _rotation_finished:
         # Note that "up" is "down". Behavior of "rotate up" is that the camera moves down toward vegetal.
         # The camera angle indeed "rotates up" around it's OWN axis while its position rotates DOWN relative to the
         # universe at the same time – continuing to point toward the center of the universe the whole time.
         # The effect is that the particles rotate upward in the field of view, i.e. the camera rotates downward.
         tf.system.camera_rotate_up()
-        _test_when_to_finish_rotation()
     
 def init_screenshots() -> None:
     """
@@ -116,18 +129,15 @@ def init_screenshots() -> None:
 def _export_screenshot(filename: str) -> None:
     """Note, in system.screenshot(), bgcolor (and decorate) args only work in windowless mode."""
     path: str = os.path.join(_image_path, filename)
-    print(f"Saving screenshot to '{path}'")
     result: int = tf.system.screenshot(path, bgcolor=[0, 0, 0])
     if result != 0:
         print(tfu.bluecolor + f"Something went wrong with screenshot export, result code = {result}" + tfu.endcolor)
 
-def save_screenshot(filename: str, show_timestep: bool = True) -> None:
-    """
-    Calling this method directly, is intended for one-off export operations *outside* of timestep events. Thus it
-    only works in windowless mode. In windowed mode, always access this through save_screenshot_repeatedly().
+def _save_screenshot(filename: str, side: str = None, show_timestep: bool = True) -> None:
+    """side = "Front", "Back", "Left", "Right".
     
     Caller provides filename (no extension). Saves as jpg. All other formats crash the app, segfault!
-    Timestep will be appended to filename unless show_timestep = False (and filename is not blank).
+    Timestep and side will be appended to filename unless show_timestep = False (and filename is not blank).
     """
     if not screenshot_export_enabled():
         return
@@ -135,14 +145,41 @@ def save_screenshot(filename: str, show_timestep: bool = True) -> None:
     suffix: str = f"Timestep = {_current_screenshot_timestep}"
     suffix += f"; Universe.time = {round(tf.Universe.time, 2)}"
     suffix += f"; phi = {round(epu.leading_edge_mean_phi(), 2)}"
-    suffix += f"; veloc.z = {round(epu.leading_edge_velocity_z(), 4)}"
     suffix += f"; {len(g.Little.items()) + len(g.LeadingEdge.items())} cells"
+    if side:
+        suffix += f" - {side}"
+        
     if not filename:
         filename = suffix
     elif show_timestep:
         filename += "; " + suffix
     
     _export_screenshot(filename + ".jpg")
+    
+def save_screenshots(filename: str) -> None:
+    """
+    Calling this method directly, is intended for one-off export operations *outside* of timestep events. Thus it
+    only works in windowless mode. In windowed mode, always access this through save_screenshot_repeatedly().
+    (At least, doesn't work *before* the sim window opens, though seems to work after it closes.)
+    """
+    if not _rotation_started:
+        _test_when_to_start_rotation()
+        
+    print(f"Saving screenshots for Timestep = {_current_screenshot_timestep};"
+          f" phi = {round(epu.leading_edge_mean_phi(), 2)}")
+
+    if cfg.windowed_mode:
+        _automated_camera_rotate()
+        _save_screenshot(filename, "Front")
+    else:
+        for side in ["Left", "Right", "Back", "Front"]:
+            _load_camera_data(side)
+            _automated_camera_rotate()
+            _save_screenshot(filename, side)
+            _save_camera_data(side)
+    
+    if _rotation_started and not _rotation_finished:
+        _test_when_to_finish_rotation()
 
 def save_screenshot_repeatedly() -> None:
     """For use inside timestep events. Keeps track of export interval, and names files accordingly.
@@ -160,13 +197,12 @@ def save_screenshot_repeatedly() -> None:
     # But also note that that first image is really simulation timestep 1, EVEN IF the simulation is
     # starting fresh. Because by the time the event gets called in Timestep 0, the physics has
     # already been integrated, and particles have moved. If need a screenshot before that, can get it
-    # (in windowless mode only) by calling a one-off save_screenshot() before calling tf.step().
+    # (in windowless mode only) by calling a one-off save_screenshots() before calling tf.step().
     
     elapsed: int = _current_screenshot_timestep - _previous_screenshot_timestep
     if elapsed % _screenshot_export_interval == 0:
         _previous_screenshot_timestep = _current_screenshot_timestep
-        _automated_camera_rotate()
-        save_screenshot("")  # just timestep as filename
+        save_screenshots("")
     
     _current_screenshot_timestep += 1
 
@@ -209,6 +245,49 @@ def screenshot_export_enabled() -> bool:
     """Convenience function. Interpret cfg.screenshot_export_interval as flag for whether export is enabled"""
     return cfg.screenshot_export_interval != 0
 
+# A dict of four CameraData objects, with keys "Front", "Back", "Left", "Right".
+# (For windowless only. In windowed mode, this data cannot be captured.)
+_camera_data: dict[str, CameraData] = {}
+
+def init_camera_data() -> None:
+    if cfg.windowed_mode:
+        tf.system.camera_view_front()
+        tf.system.camera_zoom_to(-12)
+    else:
+        # Need to render after each of these movements in order to save valid values
+        tf.system.camera_view_left()
+        tf.system.camera_zoom_to(-12)
+        _export_and_delete_a_junk_screenshot()
+        _save_camera_data("Left")
+        
+        tf.system.camera_view_right()
+        tf.system.camera_zoom_to(-12)
+        _export_and_delete_a_junk_screenshot()
+        _save_camera_data("Right")
+        
+        tf.system.camera_view_back()
+        tf.system.camera_zoom_to(-12)
+        _export_and_delete_a_junk_screenshot()
+        _save_camera_data("Back")
+        
+        tf.system.camera_view_front()
+        tf.system.camera_zoom_to(-12)
+        _export_and_delete_a_junk_screenshot()
+        _save_camera_data("Front")
+    
+def _save_camera_data(side: str) -> None:
+    """side = "Front", "Back", "Left", "Right"."""
+    center: tf.fVector3 = tf.system.camera_center()
+    rotation: tf.fQuaternion = tf.system.camera_rotation()
+    zoom: float = tf.system.camera_zoom()
+
+    _camera_data[side] = CameraData(center, rotation, zoom)
+    
+def _load_camera_data(side: str) -> None:
+    """Do this just before rendering, to make sure everything stays in sync"""
+    data: CameraData = _camera_data[side]
+    tf.system.camera_move_to(data.center, data.rotation, data.zoom)
+
 def get_state() -> dict:
     """generate state to be saved to disk
     
@@ -217,19 +296,29 @@ def get_state() -> dict:
     Other aspects of state should automatically reconstitute themselves pretty well:
     _image_path, _rotation_started
     """
-    center: tf.fVector3 = tf.system.camera_center()
-    zoom: float = tf.system.camera_zoom()
-    rotation: tf.fQuaternion = tf.system.camera_rotation()
-    angle: float = rotation.angle()
-    axis: tf.fVector3 = rotation.axis()
+    camera_json: dict = {}
+    if not cfg.windowed_mode:
+        # In windowed mode, camera data cannot be captured. (save_camera_data() fails.) So there's no
+        # way to save it to disk and recover that state later. This means, if you try to save and restore
+        # state late in the sim, after rotation has begun, all camera info will be lost.
+        for side in ["Left", "Right", "Back", "Front"]:
+            data: CameraData = _camera_data[side]
+            center: tf.fVector3 = data.center
+            zoom: float = data.zoom
+            rotation: tf.fQuaternion = data.rotation
+            angle: float = rotation.angle()
+            axis: tf.fVector3 = rotation.axis()
+            camera_json[side] = {"camera_center": center.as_list(),
+                                 "camera_zoom": zoom,
+                                 "camera_angle": angle,
+                                 "camera_axis": axis.as_list(),
+                                 }
 
     return {"previous_step": _previous_screenshot_timestep,
             "current_step": _current_screenshot_timestep,
             "rotation_finished": _rotation_finished,
-            "camera_center": center.as_list(),
-            "camera_zoom": zoom,
-            "camera_angle": angle,
-            "camera_axis": axis.as_list()}
+            "camera_data": camera_json,
+            }
 
 def set_state(d: dict) -> None:
     """Reconstitute state of module from what was saved.
@@ -242,20 +331,39 @@ def set_state(d: dict) -> None:
     _current_screenshot_timestep = d["current_step"]
     _rotation_finished = d["rotation_finished"]
     
-    camera_center: tf.fVector3 = tf.fVector3(d["camera_center"])
-    camera_zoom: float = d["camera_zoom"]
-    camera_angle: float = d["camera_angle"]
-    camera_axis: tf.fVector3 = tf.fVector3(d["camera_axis"])
-    camera_rotation: tf.fQuaternion = tf.fQuaternion.rotation(camera_angle, camera_axis)
-    tf.system.camera_move_to(camera_center, camera_rotation, camera_zoom)
+    if cfg.windowed_mode:
+        # No camera state was saved, so recover as best we can.
+        # I suppose I could calculate the position, if I were to capture and save the screenshot_timestep value
+        # when rotation started, then get the number of screenshots (hence rotations) that would have occurred
+        # since then, and from that get the new camera position. But, nah.
+        if _rotation_finished:
+            tf.system.camera_view_bottom()
+            tf.system.camera_zoom_to(-12)
+        else:
+            tf.system.camera_view_front()
+            tf.system.camera_zoom_to(-12)
+        return
     
-    # camera_move_to() doesn't actually change the camera state; it just says what it WILL move to,
+    camera_json: dict = d["camera_data"]
+    for side, side_value in camera_json.items():
+        camera_center: tf.fVector3 = tf.fVector3(side_value["camera_center"])
+        camera_zoom: float = side_value["camera_zoom"]
+        camera_angle: float = side_value["camera_angle"]
+        camera_axis: tf.fVector3 = tf.fVector3(side_value["camera_axis"])
+        camera_rotation: tf.fQuaternion = tf.fQuaternion.rotation(camera_angle, camera_axis)
+        _camera_data[side] = CameraData(camera_center, camera_rotation, camera_zoom)
+    _load_camera_data("Front")
+    
+    # _load_camera_data() doesn't actually change the camera state; it just says what it WILL move to,
     # on the subsequent render. So we need to render right now, in order get things in sync, so that
     # when the camera state is examined (see _test_when_to_finish_rotation()), it will be correct.
     _export_and_delete_a_junk_screenshot()
 
 def _export_and_delete_a_junk_screenshot() -> None:
     """The only way I know to force a render. Necessary to make recent camera changes readable."""
+    if cfg.windowed_mode:
+        return
+    
     path: str = os.path.join(_image_path, "junk.jpg")
     tf.system.screenshot(path)
     os.remove(path)
@@ -264,27 +372,29 @@ def make_movie(filename: str = None) -> None:
     if not screenshot_export_enabled():
         return
 
+    sides: list[str] = ["Front"] if cfg.windowed_mode else ["Left", "Right", "Back", "Front"]
     with os.scandir(_image_path) as dir_entries_it:
         dir_entries_chron: list[os.DirEntry] = sorted(dir_entries_it, key=lambda entry: entry.stat().st_mtime_ns)
-    image_filepaths = [entry.path
-                       for entry in dir_entries_chron
-                       if entry.name.endswith(".jpg")]
-    print(f"Assembling movie from {len(image_filepaths)} images")
-    clip = movieclip.ImageSequenceClip(image_filepaths, fps=24)
-    
-    # Save the clip using tfu.export_directory() also as the movie name, and save it to the Screenshots subdirectory
-    # (Or, if filename was provided, then __name__ == "__main__", see below.)
-    if filename is None:
-        filename = tfu.export_directory()
-    clip.write_videofile(os.path.join(_image_path, filename + ".mp4"))
-    
-    # Discard all the exported image files except the final one.
-    # (But not if there was an exception, because I may still need them.)
-    if not events.event_exception_was_thrown():
-        image_filepaths.pop()   # remove final item
-        print(f"\nRemoving {len(image_filepaths)} images...")
-        for path in image_filepaths:
-            os.remove(path)
+    for side in sides:
+        image_filepaths = [entry.path
+                           for entry in dir_entries_chron
+                           if entry.name.endswith(f"{side}.jpg")]
+        print(f"Assembling movie \"{side}\" from {len(image_filepaths)} images")
+        clip = movieclip.ImageSequenceClip(image_filepaths, fps=24)
+        
+        # Save the clip using tfu.export_directory() also as the movie name, and save it to the Screenshots subdirectory
+        # (Or, if filename was provided, then __name__ == "__main__", see below.)
+        if filename is None:
+            filename = tfu.export_directory()
+        clip.write_videofile(os.path.join(_image_path, filename + f" {side}.mp4"))
+        
+        # Discard all the exported image files except the final one.
+        # (But not if there was an exception, because I may still need them.)
+        if not events.event_exception_was_thrown():
+            image_filepaths.pop()   # remove final item
+            print(f"\nRemoving {len(image_filepaths)} images...")
+            for path in image_filepaths:
+                os.remove(path)
 
     # Notes on codec and file type:
     # .mp4 (defaults to codec "libx264"): looks okay; quality not as good as the jpg files it's made from;
